[Decomp][Player] Cohesion split: extract spell-modifier application to PlayerSpellMod.cpp
