+++ src/game/Object/Player.cpp
@@ -5550,117 +5550,6 @@ void Player::UpdateDuelFlag(time_t currTime)
     duel->opponent->duel->startTime  = currTime;
 }
 
-/**
- * @brief Applies or removes a spell modifier and notifies the client.
- *
- * @param aura The aura whose modifier should be added or removed.
- * @param apply True to apply the modifier; false to remove it.
- */
-void Player::AddSpellMod(Aura* aura, bool apply)
-{
-    Modifier const* mod = aura->GetModifier();
-    OpcodesList opcode = (mod->m_auraname == SPELL_AURA_ADD_FLAT_MODIFIER) ? SMSG_SET_FLAT_SPELL_MODIFIER : SMSG_SET_PCT_SPELL_MODIFIER;
-
-    uint32 modTypeCount = 0;    // count of mods per one mod->op
-    WorldPacket data(opcode, 4 + 4 + 1 + 1 + 4);
-    data << uint32(1);          // count of different mod->op's in packet
-    size_t writePos = data.wpos();
-    data << uint32(modTypeCount);
-    data << uint8(mod->m_miscvalue);
-    for (int eff = 0; eff < 96; ++eff)
-    {
-        uint64 _mask = 0;
-        uint32 _mask2 = 0;
-
-        if (eff < 64)
-        {
-            _mask = uint64(1) << (eff - 0);
-        }
-        else
-        {
-            _mask2 = uint32(1) << (eff - 64);
-        }
-
-        if (aura->GetAuraSpellClassMask().IsFitToFamilyMask(_mask, _mask2))
-        {
-            int32 val = 0;
-            for (AuraList::const_iterator itr = m_spellMods[mod->m_miscvalue].begin(); itr != m_spellMods[mod->m_miscvalue].end(); ++itr)
-            {
-                if ((*itr)->GetModifier()->m_auraname == mod->m_auraname && ((*itr)->GetAuraSpellClassMask().IsFitToFamilyMask(_mask, _mask2)))
-                {
-                    val += (*itr)->GetModifier()->m_amount;
-                }
-            }
-            val += apply ? mod->m_amount : -(mod->m_amount);
-            data << uint8(eff);
-            data << float(val);
-            ++modTypeCount;
-        }
-    }
-    data.put<uint32>(writePos, modTypeCount);
-    SendDirectMessage(&data);
-
-    if (apply)
-    {
-        m_spellMods[mod->m_miscvalue].push_back(aura);
-    }
-    else
-    {
-        m_spellMods[mod->m_miscvalue].remove(aura);
-    }
-}
-
-template <class T> T Player::ApplySpellMod(uint32 spellId, SpellModOp op, T& basevalue, Spell const* /*spell*/)
-{
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(spellId);
-    if (!spellInfo)
-    {
-        return 0;
-    }
-
-    int32 totalpct = 0;
-    int32 totalflat = 0;
-    for (AuraList::iterator itr = m_spellMods[op].begin(); itr != m_spellMods[op].end(); ++itr)
-    {
-        Aura* aura = *itr;
-
-        Modifier const* mod = aura->GetModifier();
-
-        if (!aura->isAffectedOnSpell(spellInfo))
-        {
-            continue;
-        }
-
-        if (mod->m_auraname == SPELL_AURA_ADD_FLAT_MODIFIER)
-        {
-            totalflat += mod->m_amount;
-        }
-        else
-        {
-            // skip percent mods for null basevalue (most important for spell mods with charges )
-            if (basevalue == T(0))
-            {
-                continue;
-            }
-
-            // special case (skip >10sec spell casts for instant cast setting)
-            if (mod->m_miscvalue == SPELLMOD_CASTING_TIME
-                    && basevalue >= T(10 * IN_MILLISECONDS) && mod->m_amount <= -100)
-                continue;
-
-            totalpct += mod->m_amount;
-        }
-    }
-
-    float diff = (float)basevalue * (float)totalpct / 100.0f + (float)totalflat;
-    basevalue = T((float)basevalue + diff);
-    return T(diff);
-}
-
-template int32 Player::ApplySpellMod<int32>(uint32 spellId, SpellModOp op, int32& basevalue, Spell const* spell);
-template uint32 Player::ApplySpellMod<uint32>(uint32 spellId, SpellModOp op, uint32& basevalue, Spell const* spell);
-template float Player::ApplySpellMod<float>(uint32 spellId, SpellModOp op, float& basevalue, Spell const* spell);
-
 // send Proficiency
 void Player::SendProficiency(ItemClass itemClass, uint32 itemSubclassMask)
 {
